+++ changelog.md
@@ -1,3 +1,8 @@
+== Handheld Piston Mod 0.3 ==
+* Updated to 1.7.10
+* Improved redstone remote to allow more than one powered block
+* Fixed rendering issues with moving block
+
 == Handheld Piston Mod 0.2 ==
 * Updated to 1.7.2
 * Fixed moving block rendering